backend/app/requirements.txt

Flask
Jinja2
Werkzeug
gunicorn
tensorflow==2.0.0
numpy
matplotlib
Pillow
